+++ keyboards/tarohayashi/undertow/lib/common_undertow.c
@@ -19,8 +19,8 @@ ut_config_t ut_config;         // eeprom保存用
 bool force_scrolling, force_cursoring, force_key_input, force_gaming, slow_mode;          // 一時的モード変更用
 uint8_t joystick_attached;     // ジョイスティックの有無
 bool joystick_initialized;
-double prev_x_0, prev_y_0, prev_x_1, prev_y_1;
-double x_accumulator, y_accumulator, h_accumulator, v_accumulator;  // 端数保存用
+float prev_x_0, prev_y_0, prev_x_1, prev_y_1;
+float x_accumulator, y_accumulator, h_accumulator, v_accumulator;  // 端数保存用
 int16_t gp29_newt, gp28_newt;              // ジョイスティックの初期値
 int16_t gp29_max, gp28_max, gp29_min, gp28_min; // ジョイスティックの最大値、最小値
 uint16_t joystick_offset_min, joystick_offset_max; // ジョイスティックの有無ョイスティックの小さい値、大きい値を無視する範囲
@@ -216,20 +216,20 @@ report_mouse_t pointing_device_task_kb(report_mouse_t mouse_report) {
 
     /* SIDE0 */
     // 数値取得、角度、反転処理
-    double rad = (double)ut_config.angle_0 * 12.0 * (M_PI / 180.0) * -1.0;
-    double x_rev_0 =  + mouse_report.x * cos(rad) - mouse_report.y * sin(rad);
-    double y_rev_0 =  + mouse_report.x * sin(rad) + mouse_report.y * cos(rad);
-    double h_rev_0 = 0.0;
-    double v_rev_0 = 0.0;
+    float rad = (float)ut_config.angle_0 * 12.0 * (M_PI / 180.0) * -1.0;
+    float x_rev_0 =  + mouse_report.x * cos(rad) - mouse_report.y * sin(rad);
+    float y_rev_0 =  + mouse_report.x * sin(rad) + mouse_report.y * cos(rad);
+    float h_rev_0 = 0.0;
+    float v_rev_0 = 0.0;
 
-    double smoothed_x_0 = prev_x_0 * SMOOTHING_FACTOR + x_rev_0 * (1.0 - SMOOTHING_FACTOR);
-    double smoothed_y_0 = prev_y_0 * SMOOTHING_FACTOR + y_rev_0 * (1.0 - SMOOTHING_FACTOR);
+    float smoothed_x_0 = prev_x_0 * SMOOTHING_FACTOR + x_rev_0 * (1.0 - SMOOTHING_FACTOR);
+    float smoothed_y_0 = prev_y_0 * SMOOTHING_FACTOR + y_rev_0 * (1.0 - SMOOTHING_FACTOR);
     prev_x_0 = smoothed_x_0;
     prev_y_0 = smoothed_y_0;
 
     // 動作量で移動量を変える
-    double movement_magnitude_0 = sqrt(smoothed_x_0 * smoothed_x_0 + smoothed_y_0 * smoothed_y_0);
-    double dynamic_multiplier_0 = 1.0 + movement_magnitude_0 / 10.0;
+    float movement_magnitude_0 = sqrt(smoothed_x_0 * smoothed_x_0 + smoothed_y_0 * smoothed_y_0);
+    float dynamic_multiplier_0 = 1.0 + movement_magnitude_0 / 10.0;
     dynamic_multiplier_0 = fmin(fmax(dynamic_multiplier_0, 0.5), 3.0);
     x_rev_0 *= SENSITIVITY_MULTIPLIER * dynamic_multiplier_0;
     y_rev_0 *= SENSITIVITY_MULTIPLIER * dynamic_multiplier_0;
@@ -300,20 +300,20 @@ report_mouse_t pointing_device_task_kb(report_mouse_t mouse_report) {
     /* SIDE1 */
     // 数値取得、角度、反転処理
     pmw33xx_report_t report = pmw33xx_read_burst(1);
-    rad = (double)ut_config.angle_1 * 12.0 * (M_PI / 180.0) * -1.0;
-    double x_rev_1 =  + report.delta_x * cos(rad) - report.delta_y * sin(rad);
-    double y_rev_1 =  + report.delta_x * sin(rad) + report.delta_y * cos(rad);
-    double h_rev_1 = 0.0;
-    double v_rev_1 = 0.0;
+    rad = (float)ut_config.angle_1 * 12.0 * (M_PI / 180.0) * -1.0;
+    float x_rev_1 =  + report.delta_x * cos(rad) - report.delta_y * sin(rad);
+    float y_rev_1 =  + report.delta_x * sin(rad) + report.delta_y * cos(rad);
+    float h_rev_1 = 0.0;
+    float v_rev_1 = 0.0;
 
-    double smoothed_x_1 = prev_x_1 * SMOOTHING_FACTOR + x_rev_1 * (1.0 - SMOOTHING_FACTOR);
-    double smoothed_y_1 = prev_y_1 * SMOOTHING_FACTOR + y_rev_1 * (1.0 - SMOOTHING_FACTOR);
+    float smoothed_x_1 = prev_x_1 * SMOOTHING_FACTOR + x_rev_1 * (1.0 - SMOOTHING_FACTOR);
+    float smoothed_y_1 = prev_y_1 * SMOOTHING_FACTOR + y_rev_1 * (1.0 - SMOOTHING_FACTOR);
     prev_x_1 = smoothed_x_1;
     prev_y_1 = smoothed_y_1;
 
     // 動作量で移動量を変える
-    double movement_magnitude_1 = sqrt(smoothed_x_1 * smoothed_x_1 + smoothed_y_1 * smoothed_y_1);
-    double dynamic_multiplier_1 = 1.0 + movement_magnitude_1 / 10.0;
+    float movement_magnitude_1 = sqrt(smoothed_x_1 * smoothed_x_1 + smoothed_y_1 * smoothed_y_1);
+    float dynamic_multiplier_1 = 1.0 + movement_magnitude_1 / 10.0;
     dynamic_multiplier_1 = fmin(fmax(dynamic_multiplier_1, 0.5), 3.0);
     x_rev_1 *= SENSITIVITY_MULTIPLIER * dynamic_multiplier_1;
     y_rev_1 *= SENSITIVITY_MULTIPLIER * dynamic_multiplier_1;
@@ -382,33 +382,33 @@ report_mouse_t pointing_device_task_kb(report_mouse_t mouse_report) {
     }
 
     /* JOYSTICK */
-    double x_rev_js = 0.0;
-    double y_rev_js = 0.0;
-    double h_rev_js = 0.0;
-    double v_rev_js = 0.0;
+    float x_rev_js = 0.0;
+    float y_rev_js = 0.0;
+    float h_rev_js = 0.0;
+    float v_rev_js = 0.0;
     if(joystick_attached != 2){
         // モード決定
-        double amp_temp;
+        float amp_temp;
         bool inv_js;
         if(joystick_attached == 0){
             cur_mode = ut_config.pd_mode_0;
-            rad = (double)ut_config.angle_0 * 12.0 * (M_PI / 180.0) * -1.0;
+            rad = (float)ut_config.angle_0 * 12.0 * (M_PI / 180.0) * -1.0;
             layer = layer_switch_get_layer(key_up_0);
             keycode_up_js = keymap_key_to_keycode(layer, key_up_0);
             keycode_left_js = keymap_key_to_keycode(layer, key_left_0);
             keycode_right_js = keymap_key_to_keycode(layer, key_right_0);
             keycode_down_js = keymap_key_to_keycode(layer, key_down_0);
-            amp_temp = 16.0 + (double)ut_config.spd_0 * 3.0;
+            amp_temp = 16.0 + (float)ut_config.spd_0 * 3.0;
             inv_js = ut_config.inv_0;
         }else{
             cur_mode = ut_config.pd_mode_1;
-            rad = (double)ut_config.angle_1 * 12.0 * (M_PI / 180.0) * -1.0;
+            rad = (float)ut_config.angle_1 * 12.0 * (M_PI / 180.0) * -1.0;
             layer = layer_switch_get_layer(key_up_1);
             keycode_up_js = keymap_key_to_keycode(layer, key_up_1);
             keycode_left_js = keymap_key_to_keycode(layer, key_left_1);
             keycode_right_js = keymap_key_to_keycode(layer, key_right_1);
             keycode_down_js = keymap_key_to_keycode(layer, key_down_1);
-            amp_temp = 16.0 + (double)ut_config.spd_1 * 3.0;
+            amp_temp = 16.0 + (float)ut_config.spd_1 * 3.0;
             inv_js = ut_config.inv_1;
         }
         if(force_cursoring){
@@ -457,8 +457,8 @@ report_mouse_t pointing_device_task_kb(report_mouse_t mouse_report) {
             gp28_val = 0;
         }
 
-        double x_val_js = ( (double)temp_x_val / JOYSTICK_DIVISOR ) * amp_temp;
-        double y_val_js = ( (double)temp_y_val / JOYSTICK_DIVISOR ) * amp_temp;
+        float x_val_js = ( (float)temp_x_val / JOYSTICK_DIVISOR ) * amp_temp;
+        float y_val_js = ( (float)temp_y_val / JOYSTICK_DIVISOR ) * amp_temp;
         // 角度補正
         x_rev_js =  + x_val_js * cos(rad) - y_val_js * sin(rad);
         y_rev_js =  + x_val_js * sin(rad) + y_val_js * cos(rad);
@@ -523,19 +523,19 @@ report_mouse_t pointing_device_task_kb(report_mouse_t mouse_report) {
             y_rev_js = 0.0;
 
             // セットアップ
-            double x_val_gp = (double)(gp29_val - gp29_newt);
-            double y_val_gp = (double)(gp28_val - gp28_newt);
+            float x_val_gp = (float)(gp29_val - gp29_newt);
+            float y_val_gp = (float)(gp28_val - gp28_newt);
             // xが正 = 右の場合
             if(gp29_val > gp29_newt){
                 // 値が小さければ0
-                if(gp29_val < (double)(gp29_newt + joystick_offset_min)){
+                if(gp29_val < (float)(gp29_newt + joystick_offset_min)){
                     x_val_gp = 0.0;
                 // 値が大きければ減らす
-                }else if(gp29_val > (double)(gp29_max - joystick_offset_max)){
+                }else if(gp29_val > (float)(gp29_max - joystick_offset_max)){
                     x_val_gp = x_val_gp - joystick_offset_max;
                 }
                 // 正規化
-                double span = (double)(gp29_max - joystick_offset_max - joystick_offset_min - gp29_newt);
+                float span = (float)(gp29_max - joystick_offset_max - joystick_offset_min - gp29_newt);
                 if(span > 0){
                     x_val_gp = x_val_gp / span * 511.0;
                 }else{
@@ -544,14 +544,14 @@ report_mouse_t pointing_device_task_kb(report_mouse_t mouse_report) {
             // xが負 = 左の場合
             }else{
                 // 値が小さければ0
-                if(gp29_val > (double)(gp29_newt - joystick_offset_min)){
+                if(gp29_val > (float)(gp29_newt - joystick_offset_min)){
                     x_val_gp = 0.0;
                 // 値が大きければ減らす
-                }else if(gp29_val < (double)(gp29_min + joystick_offset_max)){
+                }else if(gp29_val < (float)(gp29_min + joystick_offset_max)){
                     x_val_gp = x_val_gp + joystick_offset_max;
                 }
                 // 正規化
-                double span =  (double)(gp29_newt - joystick_offset_max - joystick_offset_min - gp29_min);
+                float span =  (float)(gp29_newt - joystick_offset_max - joystick_offset_min - gp29_min);
                 if (span > 0){
                     x_val_gp = x_val_gp / span * 511.0;
                 }else{
@@ -561,14 +561,14 @@ report_mouse_t pointing_device_task_kb(report_mouse_t mouse_report) {
             // yが正 = 上の場合
             if(gp28_val > gp28_newt){
                 // 値が小さければ0
-                if(gp28_val < (double)(gp28_newt + joystick_offset_min)){
+                if(gp28_val < (float)(gp28_newt + joystick_offset_min)){
                     y_val_gp = 0.0;
                 // 値が大きければ減らす
-                }else if(gp28_val > (double)(gp28_max - joystick_offset_max)){
+                }else if(gp28_val > (float)(gp28_max - joystick_offset_max)){
                     y_val_gp = y_val_gp - joystick_offset_max;
                 }
                 // 正規化
-                double span = (double)(gp28_max - joystick_offset_max - joystick_offset_min - gp28_newt);
+                float span = (float)(gp28_max - joystick_offset_max - joystick_offset_min - gp28_newt);
                 if(span > 0){
                     y_val_gp = y_val_gp / span * 511.0;
                 }else{
@@ -577,14 +577,14 @@ report_mouse_t pointing_device_task_kb(report_mouse_t mouse_report) {
             // yが正 = 下の場合
             }else{
                 // 値が小さければ0
-                if(gp28_val > (double)(gp28_newt - joystick_offset_min)){
+                if(gp28_val > (float)(gp28_newt - joystick_offset_min)){
                     y_val_gp = 0.0;
                 // 値が大きければ減らす
-                }else if(gp28_val < (double)(gp28_min + joystick_offset_max)){
+                }else if(gp28_val < (float)(gp28_min + joystick_offset_max)){
                     y_val_gp = y_val_gp + joystick_offset_max;
                 }
                 // 正規化
-                double span =  (double)(gp28_newt - joystick_offset_max - joystick_offset_min - gp28_min);
+                float span =  (float)(gp28_newt - joystick_offset_max - joystick_offset_min - gp28_min);
                 if (span > 0){
                     y_val_gp = y_val_gp / span * 511.0;
                 }else{
@@ -593,8 +593,8 @@ report_mouse_t pointing_device_task_kb(report_mouse_t mouse_report) {
             }
 
             // 角度の修正
-            double x_rev_gp =  + x_val_gp * cos(rad) - y_val_gp * sin(rad);
-            double y_rev_gp =  + x_val_gp * sin(rad) + y_val_gp * cos(rad);
+            float x_rev_gp =  + x_val_gp * cos(rad) - y_val_gp * sin(rad);
+            float y_rev_gp =  + x_val_gp * sin(rad) + y_val_gp * cos(rad);
 
             // x軸反転処理f
             if(inv_js){
